+++ src/qcm/solver/solver_qnn.py
@@ -1,5 +1,3 @@
-# model/models.py
-
 import torch
 import torch.nn as nn
 from torch.utils.data import DataLoader
@@ -9,108 +7,34 @@ import time
 import logging
 import numpy as np
 
-from ..components.encoder.pcam import PCAM as PCAMBackbone
-from ..components.encoder.tcga import TCGA as TCGABackbone
+class SolverQNN:
 
-from ..components.qnn.classical import ClassicalHead
-from ..components.qnn.amplitude_encoding import QuantumHeadAmplitude
-from ..components.qnn.angle_encoding import QuantumHeadAngle
-
-
-class HybridClassifier(nn.Module):
-    def __init__(self, config: dict, use_quantum: bool = False, num_quantum_layers: int = 1):
-        super().__init__()
-        model_cfg = config['model']
-        dataset_type = config['dataset_type']
-        self.num_quantum_layers = num_quantum_layers
+    def __init__(self, 
+                 model: nn.Module, 
+                 dataloader: DataLoader,
+                 optimizer: torch.optim.Optimizer,
+                 scheduler,
+                 config: dict,
+                 writer: SummaryWriter,
+                 device: torch.device,
+                 logger: logging.Logger
+                 ):
         
-        # 1. Select Backbone
-        if dataset_type == 'pcam':
-            self.backbone = PCAMBackbone(
-                latent_dim=model_cfg['latent_dim'], 
-                filters=model_cfg['pcam_filters']
-            )
-            num_classes = 1
-        elif dataset_type == 'tcga':
-            self.backbone = TCGABackbone(
-                input_dim=768, 
-                latent_dim=model_cfg['latent_dim']
-            )
-            num_classes = model_cfg['num_classes']
-        else:
-            raise ValueError(f"Unsupported dataset_type: {dataset_type}")
+        self.model = model
+        self.dataloader = dataloader
+        self.optimizer = optimizer
+        self.scheduler = scheduler
+        self.config = config
+        self.writer = writer
+        self.device = device
+        self.logger = logger
 
-        # 2. Select Head
-        if use_quantum:
-            head_type = model_cfg.get('quantum_head_type', 'amplitude')
-            n_qubits = model_cfg['n_qubits']
-            
-            if head_type == 'amplitude':
-                latent_dim = 2**n_qubits
-                if dataset_type == 'tcga':
-                     self.backbone = TCGABackbone(input_dim=768, latent_dim=latent_dim)
-                else:
-                     self.backbone = PCAMBackbone(latent_dim=latent_dim, filters=model_cfg['pcam_filters'])
-
-                self.head = QuantumHeadAmplitude(
-                    n_qubits=n_qubits,
-                    num_classes=num_classes,
-                    n_layers=self.num_quantum_layers,
-                    entangling_layer=model_cfg['entangling_layer']
-                )
-            elif head_type == 'angle':
-                latent_dim = n_qubits
-                if dataset_type == 'tcga':
-                     self.backbone = TCGABackbone(input_dim=768, latent_dim=latent_dim)
-                else:
-                     self.backbone = PCAMBackbone(latent_dim=latent_dim, filters=model_cfg['pcam_filters'])
-
-                self.head = QuantumHeadAngle(
-                    n_qubits=n_qubits,
-                    num_classes=num_classes,
-                    n_layers=self.num_quantum_layers,
-                    entangling_layer=model_cfg['entangling_layer']
-                )
-            else:
-                raise ValueError(f"Unknown quantum_head_type: {head_type}")
-        else:
-            latent_dim = model_cfg['latent_dim']
-            self.head = ClassicalHead(latent_dim, num_classes)
-            
-    def forward(self, x: torch.Tensor) -> torch.Tensor:
-        """forward function of the module.
-
-        Args:
-            x (torch.Tensor): input data
-
-        Returns:
-            torch.Tensor: output of the module
-        """
-        x = self.backbone(x)
-        return self.head(x)
     
-    def train_epoch(self, 
-                    dataloader: DataLoader, 
-                    criterion: nn.Module, 
-                    optimizer: torch.optim.Optimizer,
-                    scheduler, 
-                    epoch: int, 
-                    config: dict, 
-                    writer: SummaryWriter,
-                    device: torch.device,
-                    logger: logging.logger):
+    def train_epoch(self, epoch: int):
         """_summary_
 
         Args:
-            dataloader (DataLoader): _description_
-            criterion (nn.Module): _description_
-            optimizer (torch.optim.Optimizer): _description_
-            scheduler (_type_): _description_
             epoch (int): _description_
-            config (dict): _description_
-            writer (SummaryWriter): _description_
-            device (torch.device): _description_
-            logger (logging.logger): _description_
 
         Returns:
             _type_: _description_
