+++ src/qcm/utils/functions.py
@@ -1,5 +1,6 @@
 from pathlib import Path
 import logging
+import yaml
 from torch.utils.data import DataLoader
 from sklearn.manifold import TSNE
 import matplotlib.pyplot as plt
@@ -59,3 +60,61 @@ def visualize_latents(model: torch.nn.Module, dataloader: DataLoader, save_path:
     plt.title("t-SNE of Latent Space")
     plt.savefig(save_path)
     plt.close()
+
+def plot_all_runs_losses(all_runs_losses: dict, dataset_type: str, timestamp: str, save_path: Path):
+    """
+    Plots the training and validation loss curves for all configurations (0 to max_num_of_quantum layers).
+    :param all_runs_losses: Dictionary {num_quantum_layers: (train_losses, val_losses)}
+    :param dataset_type: Name of the dataset for the title.
+    :param timestamp: Timestamp string for unique saving.
+    :param save_path: Path to save the plot.
+    """
+    if not all_runs_losses:
+        logger.warning("No losses recorded to plot.")
+        return
+
+    logger.info(f"Generating aggregate loss plot and saving to {save_path}")
+
+    num_epochs = len(next(iter(all_runs_losses.values()))[0])
+    epochs = range(1, num_epochs + 1)
+
+    fig, ax = plt.subplots(figsize=(12, 6))
+    colors = plt.cm.viridis(np.linspace(0, 1, len(all_runs_losses)))
+    sorted_keys = sorted(all_runs_losses.keys())
+
+    for i, n_layers in enumerate(sorted_keys):
+        train_losses, val_losses = all_runs_losses[n_layers]
+        label_base = f'{n_layers} Q Layers'
+        if n_layers == 0:
+            label_base = 'Classical (0 Q Layers)'
+
+        ax.plot(epochs, train_losses,
+                label=f'{label_base} - Train Loss',
+                linestyle='-',
+                linewidth=1.5,
+                color=colors[i])
+
+        ax.plot(epochs, val_losses,
+                label=f'{label_base} - Val Loss',
+                linestyle='--',
+                linewidth=2.0,
+                color=colors[i])
+
+    ax.set_title(f'Hybrid Quantum-Classical Model Performance on {dataset_type.upper()} Dataset', fontsize=16)
+    ax.set_xlabel('Epoch', fontsize=12)
+    ax.set_ylabel('Loss', fontsize=12)
+    ax.legend(loc='upper right', ncol=2, fontsize=8)
+    ax.grid(True, which='both', linestyle='-')
+
+    try:
+        plt.savefig(save_path)
+        logger.info(f"Aggregate loss plot successfully saved to: {save_path}")
+    except Exception as e:
+        logger.error(f"Error saving plot: {e}")
+    finally:
+        plt.close(fig)
+
+def load_config(path: str) -> dict:
+    logger.debug(f"Loading config from: {path}")
+    with open(path, 'r') as f:
+        return yaml.safe_load(f)
